adding views for tradenote, industry, contactperson, company

+++ src/main/java/pl/matkan/wholesaler/controller/CompanyController.java
@@ -4,18 +4,19 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import pl.matkan.wholesaler.model.Company;
-import pl.matkan.wholesaler.service.CompanyService;
+import pl.matkan.wholesaler.service.impl.CompanyServiceImpl;
 
 
 import java.util.List;
 import java.util.Optional;
+
 @RestController
 @RequestMapping(value = "/companies")
-
 public class CompanyController {
-    private final CompanyService companyService;
-    public CompanyController(CompanyService companyService) {
-        this.companyService= companyService;
+    private final CompanyServiceImpl companyService;
+
+    public CompanyController(CompanyServiceImpl companyService) {
+        this.companyService = companyService;
     }
 
     /* READ */
@@ -29,12 +30,12 @@ public class CompanyController {
     public List<Company> getAll() {
         return companyService.findAll();
     }
+
     /* CREATE */
     @PostMapping()
     public ResponseEntity<Company> createOne(@RequestBody Company one) {
-        //System.out.println("co przyszlo:" + one.toString());
-        Company roleOne = companyService.create(one);
-        return new ResponseEntity<>(roleOne, HttpStatus.CREATED);
+        Company companyOne = companyService.create(one);
+        return new ResponseEntity<>(companyOne, HttpStatus.CREATED);
     }
 
     /* UPDATE */
@@ -50,8 +51,14 @@ public class CompanyController {
     /* DELETE */
     @DeleteMapping("/{id}")
     @ResponseStatus(code = HttpStatus.NO_CONTENT)
-    public void deleteOne(@PathVariable("id") Long id) {
-        companyService.deleteById(id);
+    public ResponseEntity<Long> deleteOne(@PathVariable("id") Long id) {
+        if (companyService.existsById(id)) {
+            companyService.deleteById(id);
+            return new ResponseEntity<>(id, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
+
     }
 
 }
